chore: attempt to fix test
