+++ ui/apps/search/src/components/routes/index.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Navigate, Route, Routes } from 'react-router-d
 import { RootComponentProps } from '@akashaorg/typings/ui';
 import { useGetLogin } from '@akashaorg/ui-awf-hooks';
 import routes, { ONBOARDING, RESULTS, SETTINGS } from '../../routes';
-import NewSearchPage from './new-search-page';
+import SearchPage from './search-page';
 import OnboardingPage from './onboarding-page';
 import SettingsPage from './settings-page';
 import Box from '@akashaorg/design-system-core/lib/components/Box';
@@ -24,7 +24,7 @@ const AppRoutes: React.FC<RootComponentProps> = props => {
             <Route
               path=":searchKeyword"
               element={
-                <NewSearchPage
+                <SearchPage
                   {...props}
                   showLoginModal={showLoginModal}
                   loginState={loginQuery.data}
@@ -34,7 +34,7 @@ const AppRoutes: React.FC<RootComponentProps> = props => {
             <Route
               path=""
               element={
-                <NewSearchPage
+                <SearchPage
                   {...props}
                   showLoginModal={showLoginModal}
                   loginState={loginQuery.data}
