+++ ui/apps/search/src/components/routes/search-page.tsx
@@ -1,555 +1,597 @@
-// import * as React from 'react';
-// import { useParams } from 'react-router-dom';
-// import { useTranslation } from 'react-i18next';
-// import DS from '@akashaorg/design-system';
+import * as React from 'react';
+import { useParams } from 'react-router-dom';
+import { useTranslation } from 'react-i18next';
+import {
+  RootComponentProps,
+  IProfileData,
+  IEntryData,
+  ITag,
+  EntityTypes,
+  ModalNavigationOptions,
+  AnalyticsCategories,
+} from '@akashaorg/typings/ui';
 
-// import {
-//   RootComponentProps,
-//   IProfileData,
-//   IEntryData,
-//   ITag,
-//   EntityTypes,
-//   ModalNavigationOptions,
-//   AnalyticsCategories,
-// } from '@akashaorg/typings/ui';
+import { ILocale } from '@akashaorg/design-system/src/utils/time';
+import routes, { SETTINGS } from '../../routes';
 
-// import { ILocale } from '@akashaorg/design-system/src/utils/time';
+import {
+  useTagSubscriptions,
+  useToggleTagSubscription,
+  useIsFollowingMultiple,
+  useFollow,
+  useUnfollow,
+  useSearchPosts,
+  useSearchProfiles,
+  useSearchComments,
+  useSearchTags,
+  LoginState,
+  useEntryNavigation,
+  useAnalytics,
+} from '@akashaorg/ui-awf-hooks';
 
-// import {
-//   useTagSubscriptions,
-//   useToggleTagSubscription,
-//   useIsFollowingMultiple,
-//   useFollow,
-//   useUnfollow,
-//   useSearchPosts,
-//   useSearchProfiles,
-//   useSearchComments,
-//   useSearchTags,
-//   LoginState,
-//   useEntryNavigation,
-//   useAnalytics,
-// } from '@akashaorg/ui-awf-hooks';
+import { SearchTagsResult } from '@akashaorg/typings/sdk/graphql-types';
 
-// import { SearchTagsResult } from '@akashaorg/typings/sdk/graphql-types';
+import EntryCardRenderer from './entry-renderer';
+import BasicCardBox from '@akashaorg/design-system-core/lib/components/BasicCardBox';
+import Box from '@akashaorg/design-system-core/lib/components/Box';
+import Divider from '@akashaorg/design-system-core/lib/components/Divider';
+import InfoCard from '@akashaorg/design-system-components/lib/components/InfoCard';
+import ProfileSearchCard from '@akashaorg/design-system-components/lib/components/ProfileSearchCard';
+import SearchDefaultCard from '@akashaorg/design-system-components/lib/components/SearchDefaultCard';
+import SearchStartCard from '@akashaorg/design-system-components/lib/components/SearchStartCard';
+import SearchAppFilter from '@akashaorg/design-system-components/lib/components/SearchAppFilter';
+import SeventyFivePercentSpinner from '@akashaorg/design-system-core/lib/components/SeventyFivePercentSpinner';
+import SwitchCard from '@akashaorg/design-system-components/lib/components/SwitchCard';
+import TagSearchCard from '@akashaorg/design-system-components/lib/components/TagSearchCard';
+import Text from '@akashaorg/design-system-core/lib/components/Text';
+import { DropdownMenuItemGroupType } from '@akashaorg/design-system-components/lib/components/SearchAppDropdownFilter';
+import SearchResultCount from './search-result-count';
+import useIntersectionObserver from '@akashaorg/design-system-core/lib/utils/intersection-observer';
 
-// import EntryCardRenderer from './entry-renderer';
+export enum ButtonValues {
+  CONTENT = 'Content',
+  PEOPLE = 'People',
+  TAGS = 'Tags',
+}
 
-// const {
-//   Box,
-//   Spinner,
-//   ProfileSearchCard,
-//   TagSearchCard,
-//   TabsToolbar,
-//   StyledSwitchCardButton,
-//   SearchStartCard,
-//   InfoCard,
-//   TAB_TOOLBAR_TYPE,
-//   useIntersectionObserver,
-// } = DS;
+interface SearchPageProps extends RootComponentProps {
+  onError?: (err: Error) => void;
+  loginState: LoginState;
+  showLoginModal: (redirectTo?: { modal: ModalNavigationOptions }) => void;
+}
 
-// export enum ButtonValues {
-//   ALL = 'All',
-//   PEOPLE = 'People',
-//   TOPICS = 'Topics',
-//   POSTS = 'Posts',
-//   REPLIES = 'Replies',
-// }
+type DataResponse = SearchTagsResult | IEntryData;
 
-// interface SearchPageProps extends RootComponentProps {
-//   onError?: (err: Error) => void;
-//   loginState: LoginState;
-//   showLoginModal: (redirectTo?: { modal: ModalNavigationOptions }) => void;
-// }
+const initSearchState = {
+  [ButtonValues.CONTENT]: { page: 1, results: [], done: false, isLoading: false },
+  [ButtonValues.PEOPLE]: { page: 1, results: [], done: false, isLoading: false },
+  [ButtonValues.TAGS]: { page: 1, results: [], done: false, isLoading: false },
+};
 
-// type DataResponse = SearchTagsResult | IEntryData;
+const SearchPage: React.FC<SearchPageProps> = props => {
+  const { singleSpa, loginState, showLoginModal } = props;
+  const { searchKeyword = '' } = useParams<{ searchKeyword: string }>();
+  const [searchState, setSearchState] = React.useState(initSearchState);
+  const [activeButton, setActiveButton] = React.useState<ButtonValues>(ButtonValues.CONTENT);
 
-// const initSearchState = {
-//   [ButtonValues.TOPICS]: { page: 1, results: [], done: false, isLoading: false },
-//   [ButtonValues.POSTS]: { page: 1, results: [], done: false, isLoading: false },
-//   [ButtonValues.PEOPLE]: { page: 1, results: [], done: false, isLoading: false },
-//   [ButtonValues.REPLIES]: { page: 1, results: [], done: false, isLoading: false },
-// };
+  const [analyticsActions] = useAnalytics();
+  const { t, i18n } = useTranslation('app-search');
+  const locale = (i18n.languages[0] || 'en') as ILocale;
 
-// const SearchPage: React.FC<SearchPageProps> = props => {
-//   const { singleSpa, loginState, showLoginModal } = props;
-//   const { searchKeyword = '' } = useParams<{ searchKeyword: string }>();
-//   const [searchState, setSearchState] = React.useState(initSearchState);
-//   const [activeButton, setActiveButton] = React.useState<ButtonValues>(ButtonValues.ALL);
+  const tagSubscriptionsReq = useTagSubscriptions(loginState?.isReady && loginState?.ethAddress);
+  const tagSubscriptionsState = tagSubscriptionsReq.data;
 
-//   const [analyticsActions] = useAnalytics();
-//   const { t, i18n } = useTranslation('app-search');
-//   const locale = (i18n.languages[0] || 'en') as ILocale;
+  const toggleTagSubscriptionReq = useToggleTagSubscription();
 
-//   const tagSubscriptionsReq = useTagSubscriptions(loginState?.isReady && loginState?.ethAddress);
-//   const tagSubscriptionsState = tagSubscriptionsReq.data;
+  const navigateTo = props.plugins['@akashaorg/app-routing']?.routing?.navigateTo;
+  const handleEntryNavigation = useEntryNavigation(navigateTo);
 
-//   const toggleTagSubscriptionReq = useToggleTagSubscription();
+  // const isAllTabActive = React.useMemo(() => activeButton === ButtonValues.CONTENT, [activeButton]);
 
-//   const navigateTo = props.plugins['@akashaorg/app-routing']?.routing?.navigateTo;
-//   const handleEntryNavigation = useEntryNavigation(navigateTo);
+  const dropdownMenuItems: DropdownMenuItemGroupType[] = [
+    {
+      id: '00',
+      title: t('All'),
+      type: 'opt',
+    },
+    {
+      id: '11',
+      title: 'Antenna',
+      type: 'optgroup',
+      children: [
+        { id: '1', title: 'Beams' },
+        { id: '2', title: 'Reflection' },
+      ],
+    },
+    {
+      id: '21',
+      title: 'Akashaverse',
+      type: 'optgroup',
+      children: [
+        { id: '4', title: 'Apps' },
+        { id: '5', title: 'Widgets' },
+        { id: '6', title: 'Plugins' },
+      ],
+    },
+  ];
 
-//   const isAllTabActive = React.useMemo(() => activeButton === ButtonValues.ALL, [activeButton]);
+  const [selected, setSelected] = React.useState<DropdownMenuItemGroupType | null>(
+    dropdownMenuItems[0],
+  );
 
-//   const updateSearchState = (
-//     type: Exclude<ButtonValues, ButtonValues.ALL>,
-//     data: Array<DataResponse & { delisted?: boolean }>,
-//   ) => {
-//     if (!data || !data.length) {
-//       setSearchState(prevState => ({
-//         ...prevState,
-//         [type]: { ...prevState[type], done: true, isLoading: false },
-//       }));
-//     }
-//     setSearchState(prevState => ({
-//       ...prevState,
-//       [type]: {
-//         ...prevState[type],
-//         results: [...prevState[type].results, ...data.filter(_ => !_.delisted)],
-//         // topics edge case because it only fetches once
-//         done: type === ButtonValues.TOPICS || prevState[type].done,
-//         isLoading: false,
-//       },
-//     }));
-//   };
+  const updateSearchState = (
+    // type: Exclude<ButtonValues, ButtonValues.PEOPLE>,
+    type: ButtonValues,
+    data: Array<DataResponse & { delisted?: boolean }>,
+  ) => {
+    if (!data || !data.length) {
+      setSearchState(prevState => ({
+        ...prevState,
+        [type]: { ...prevState[type], done: true, isLoading: false },
+      }));
+    }
+    setSearchState(prevState => ({
+      ...prevState,
+      [type]: {
+        ...prevState[type],
+        results: [...prevState[type].results, ...data.filter(_ => !_.delisted)],
+        // tags edge case because it only fetches once
+        done: type === ButtonValues.TAGS || prevState[type].done,
+        isLoading: false,
+      },
+    }));
+  };
 
-//   const handleLoadMore = () => {
-//     if (isAllTabActive) return;
-//     const { done, isLoading } = searchState[activeButton];
-//     if (done || isLoading) return;
+  const handleLoadMore = () => {
+    // if (isAllTabActive) return;
+    const { done, isLoading } = searchState[activeButton];
+    if (done || isLoading) return;
 
-//     setSearchState(prevState => ({
-//       ...prevState,
-//       [activeButton]: {
-//         ...prevState[activeButton],
-//         page: prevState[activeButton].page + 1,
-//         isLoading: true,
-//       },
-//     }));
-//   };
-//   const loadmoreRef = React.createRef<HTMLDivElement>();
+    setSearchState(prevState => ({
+      ...prevState,
+      [activeButton]: {
+        ...prevState[activeButton],
+        page: prevState[activeButton].page + 1,
+        isLoading: true,
+      },
+    }));
+  };
+  const loadmoreRef = React.createRef<HTMLDivElement>();
 
-//   useIntersectionObserver({
-//     target: loadmoreRef,
-//     onIntersect: handleLoadMore,
-//     threshold: 0,
-//   });
+  useIntersectionObserver({
+    target: loadmoreRef,
+    onIntersect: handleLoadMore,
+    threshold: 0,
+  });
 
-//   const getSearchStateForTab = (tab: Exclude<ButtonValues, ButtonValues.ALL>) => {
-//     if (activeButton === ButtonValues.ALL) {
-//       return searchState[tab].results.slice(0, 4);
-//     }
-//     return searchState[tab].results;
-//   };
+  const getSearchStateForTab = (tab: ButtonValues) => {
+    // if (activeButton === ButtonValues.CONTENT) {
+    //   return searchState[tab].results.slice(0, 4);
+    // }
+    return searchState[tab].results;
+  };
 
-//   React.useEffect(() => {
-//     if (isAllTabActive) {
-//       return setSearchState(initSearchState);
-//     }
-//     setSearchState({
-//       ...initSearchState,
-//       [activeButton]: { ...initSearchState[activeButton], isLoading: true },
-//     });
-//     // eslint-disable-next-line react-hooks/exhaustive-deps
-//   }, [searchKeyword]);
+  React.useEffect(() => {
+    // if (isAllTabActive) {
+    //   return setSearchState(initSearchState);
+    // }
+    setSearchState({
+      ...initSearchState,
+      [activeButton]: { ...initSearchState[activeButton], isLoading: true },
+    });
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [searchKeyword]);
 
-//   const searchProfilesReq = useSearchProfiles(
-//     decodeURIComponent(searchKeyword),
-//     searchState[ButtonValues.PEOPLE].page,
-//     loginState?.pubKey,
-//     loginState?.fromCache,
-//   );
-//   const searchProfilesState = getSearchStateForTab(ButtonValues.PEOPLE);
+  const searchProfilesReq = useSearchProfiles(
+    decodeURIComponent(searchKeyword),
+    searchState[ButtonValues.PEOPLE].page,
+    loginState?.pubKey,
+    loginState?.fromCache,
+  );
+  const searchProfilesState = getSearchStateForTab(ButtonValues.PEOPLE);
 
-//   const searchPostsReq = useSearchPosts(
-//     decodeURIComponent(searchKeyword),
-//     searchState[ButtonValues.POSTS].page,
-//     loginState?.pubKey,
-//     loginState?.fromCache,
-//   );
-//   const searchPostsState = getSearchStateForTab(ButtonValues.POSTS);
+  const searchBeamsReq = useSearchPosts(
+    decodeURIComponent(searchKeyword),
+    searchState[ButtonValues.CONTENT].page,
+    loginState?.pubKey,
+    loginState?.fromCache,
+  );
+  const searchBeamsState = getSearchStateForTab(ButtonValues.CONTENT);
 
-//   const searchCommentsReq = useSearchComments(
-//     decodeURIComponent(searchKeyword),
-//     searchState[ButtonValues.REPLIES].page,
-//     loginState?.pubKey,
-//     loginState?.fromCache,
-//   );
-//   const searchCommentsState = searchState[ButtonValues.REPLIES].results;
+  const searchCommentsReq = useSearchComments(
+    decodeURIComponent(searchKeyword),
+    searchState[ButtonValues.CONTENT].page,
+    loginState?.pubKey,
+    loginState?.fromCache,
+  );
+  // const searchReflectionsState = searchState[ButtonValues.CONTENT].results;
 
-//   const searchTagsReq = useSearchTags(decodeURIComponent(searchKeyword));
-//   const searchTagsState = getSearchStateForTab(ButtonValues.TOPICS);
+  const searchTagsReq = useSearchTags(decodeURIComponent(searchKeyword));
+  const searchTagsState = getSearchStateForTab(ButtonValues.TAGS);
 
-//   React.useEffect(() => {
-//     if (searchPostsReq.isFetched) updateSearchState(ButtonValues.POSTS, searchPostsReq.data);
-//   }, [searchPostsReq.data, searchPostsReq.isFetched]);
+  React.useEffect(() => {
+    if (searchBeamsReq.isFetched) {
+      updateSearchState(ButtonValues.CONTENT, searchBeamsReq.data);
+    }
+  }, [searchBeamsReq.data, searchBeamsReq.isFetched]);
 
-//   React.useEffect(() => {
-//     if (searchCommentsReq.isFetched) {
-//       updateSearchState(ButtonValues.REPLIES, searchCommentsReq.data);
-//     }
-//   }, [searchCommentsReq.data, searchCommentsReq.isFetched]);
+  React.useEffect(() => {
+    if (searchCommentsReq.isFetched) {
+      updateSearchState(ButtonValues.CONTENT, searchCommentsReq.data);
+    }
+  }, [searchCommentsReq.data, searchCommentsReq.isFetched]);
 
-//   React.useEffect(() => {
-//     if (searchProfilesReq.isFetched) {
-//       updateSearchState(ButtonValues.PEOPLE, searchProfilesReq.data);
-//     }
-//   }, [searchProfilesReq.data, searchProfilesReq.isFetched]);
+  React.useEffect(() => {
+    if (searchProfilesReq.isFetched) {
+      updateSearchState(ButtonValues.PEOPLE, searchProfilesReq.data);
+    }
+  }, [searchProfilesReq.data, searchProfilesReq.isFetched]);
 
-//   React.useEffect(() => {
-//     if (searchTagsReq.isFetched) updateSearchState(ButtonValues.TOPICS, searchTagsReq.data);
-//   }, [searchTagsReq.data, searchTagsReq.isFetched]);
+  React.useEffect(() => {
+    if (searchTagsReq.isFetched) updateSearchState(ButtonValues.TAGS, searchTagsReq.data);
+  }, [searchTagsReq.data, searchTagsReq.isFetched]);
 
-//   const followPubKeyArr = searchProfilesState?.map(profile => profile.pubKey);
-//   const isFollowingMultipleReq = useIsFollowingMultiple(loginState?.pubKey, followPubKeyArr);
-//   const followedProfiles = isFollowingMultipleReq.data;
-//   const followReq = useFollow();
-//   const unfollowReq = useUnfollow();
+  const followPubKeyArr = searchProfilesState?.map(profile => profile.pubKey);
+  const isFollowingMultipleReq = useIsFollowingMultiple(loginState?.pubKey, followPubKeyArr);
+  const followedProfiles = isFollowingMultipleReq.data;
+  const followReq = useFollow();
+  const unfollowReq = useUnfollow();
 
-//   const handleTagSubscribe = (subscribe: boolean) => (tagName: string) => {
-//     if (!loginState?.ethAddress) {
-//       showLoginModal();
-//       return;
-//     }
-//     analyticsActions.trackEvent({
-//       category: AnalyticsCategories.FILTER_SEARCH,
-//       action: subscribe ? 'Trending Topic Subscribed' : 'Trending Topic Unsubscribed',
-//     });
-//     toggleTagSubscriptionReq.mutate(tagName);
-//   };
+  const handleTagSubscribe = (subscribe: boolean) => (tagName: string) => {
+    if (!loginState?.ethAddress) {
+      showLoginModal();
+      return;
+    }
+    analyticsActions.trackEvent({
+      category: AnalyticsCategories.FILTER_SEARCH,
+      action: subscribe ? 'Trending Topic Subscribed' : 'Trending Topic Unsubscribed',
+    });
+    toggleTagSubscriptionReq.mutate(tagName);
+  };
 
-//   const handleProfileClick = (pubKey: string) => {
-//     navigateTo?.({
-//       appName: '@akashaorg/app-profile',
-//       getNavigationUrl: navRoutes => `${navRoutes.rootRoute}/${pubKey}`,
-//     });
-//   };
-//   const handleFollowProfile = (pubKey: string) => {
-//     if (!loginState?.ethAddress) {
-//       showLoginModal();
-//       return;
-//     }
-//     analyticsActions.trackEvent({
-//       category: AnalyticsCategories.FILTER_SEARCH,
-//       action: 'Trending People Followed',
-//     });
-//     followReq.mutate(pubKey);
-//   };
+  const handleProfileClick = (pubKey: string) => {
+    navigateTo?.({
+      appName: '@akashaorg/app-profile',
+      getNavigationUrl: navRoutes => `${navRoutes.rootRoute}/${pubKey}`,
+    });
+  };
+  const handleFollowProfile = (pubKey: string) => {
+    if (!loginState?.ethAddress) {
+      showLoginModal();
+      return;
+    }
+    analyticsActions.trackEvent({
+      category: AnalyticsCategories.FILTER_SEARCH,
+      action: 'Trending People Followed',
+    });
+    followReq.mutate(pubKey);
+  };
 
-//   const handleSearch = (inputValue: string) => {
-//     const trimmedValue = inputValue.trim();
-//     if (!trimmedValue) return;
-//     const encodedSearchKey = encodeURIComponent(trimmedValue);
-//     props.plugins['@akashaorg/app-routing']?.routing?.navigateTo?.({
-//       appName: '@akashaorg/app-search',
-//       getNavigationUrl: routes => `${routes.Results}/${encodedSearchKey}`,
-//     });
-//   };
+  const handleSearch = (inputValue: string) => {
+    const trimmedValue = inputValue.trim();
+    if (!trimmedValue) return;
+    const encodedSearchKey = encodeURIComponent(trimmedValue);
+    props.plugins['@akashaorg/app-routing']?.routing?.navigateTo?.({
+      appName: '@akashaorg/app-search',
+      getNavigationUrl: routes => `${routes.Results}/${encodedSearchKey}`,
+    });
+  };
 
-//   const handleAvatarClick = (ev: React.MouseEvent<HTMLDivElement>, authorEth: string) => {
-//     ev.preventDefault();
-//     navigateTo?.({
-//       appName: '@akashaorg/app-profile',
-//       getNavigationUrl: navRoutes => `${navRoutes.rootRoute}/${authorEth}`,
-//     });
-//   };
+  const handleAvatarClick = (ev: React.MouseEvent<HTMLDivElement>, authorEth: string) => {
+    ev.preventDefault();
+    navigateTo?.({
+      appName: '@akashaorg/app-profile',
+      getNavigationUrl: navRoutes => `${navRoutes.rootRoute}/${authorEth}`,
+    });
+  };
 
-//   const handleMentionClick = (profileEthAddress: string) => {
-//     navigateTo?.({
-//       appName: '@akashaorg/app-profile',
-//       getNavigationUrl: navRoutes => `${navRoutes.rootRoute}/${profileEthAddress}`,
-//     });
-//   };
+  const handleMentionClick = (profileEthAddress: string) => {
+    navigateTo?.({
+      appName: '@akashaorg/app-profile',
+      getNavigationUrl: navRoutes => `${navRoutes.rootRoute}/${profileEthAddress}`,
+    });
+  };
 
-//   const handleUnfollowProfile = (pubKey: string) => {
-//     if (!loginState?.ethAddress) {
-//       showLoginModal();
-//       return;
-//     }
-//     analyticsActions.trackEvent({
-//       category: AnalyticsCategories.FILTER_SEARCH,
-//       action: 'Trending People Unfollowed',
-//     });
-//     unfollowReq.mutate(pubKey);
-//   };
+  const handleUnfollowProfile = (pubKey: string) => {
+    if (!loginState?.ethAddress) {
+      showLoginModal();
+      return;
+    }
+    analyticsActions.trackEvent({
+      category: AnalyticsCategories.FILTER_SEARCH,
+      action: 'Trending People Unfollowed',
+    });
+    unfollowReq.mutate(pubKey);
+  };
 
-//   const handleTagClick = (name: string) => {
-//     navigateTo?.({
-//       appName: '@akashaorg/app-akasha-integration',
-//       getNavigationUrl: navRoutes => `${navRoutes.Tags}/${name}`,
-//     });
-//   };
+  const handleTagClick = (name: string) => {
+    navigateTo?.({
+      appName: '@akashaorg/app-akasha-integration',
+      getNavigationUrl: navRoutes => `${navRoutes.Tags}/${name}`,
+    });
+  };
 
-//   // repost related
-//   const handleRepost = (_withComment: boolean, itemId: string) => {
-//     analyticsActions.trackEvent({
-//       category: AnalyticsCategories.POST,
-//       action: 'Repost Clicked',
-//     });
-//     if (!loginState?.ethAddress) {
-//       showLoginModal();
-//       return;
-//     } else {
-//       navigateTo?.({
-//         appName: '@akashaorg/app-akasha-integration',
-//         getNavigationUrl: () => `/feed?repost=${itemId}`,
-//       });
-//     }
-//   };
+  // // repost related
+  const handleRebeam = (_withComment: boolean, itemId: string) => {
+    analyticsActions.trackEvent({
+      category: AnalyticsCategories.POST,
+      action: 'Repost Clicked',
+    });
+    if (!loginState?.ethAddress) {
+      showLoginModal();
+      return;
+    } else {
+      navigateTo?.({
+        appName: '@akashaorg/app-akasha-integration',
+        getNavigationUrl: () => `/feed?repost=${itemId}`,
+      });
+    }
+  };
 
-//   const emptySearchState =
-//     searchProfilesState?.length === 0 &&
-//     searchPostsState?.length === 0 &&
-//     searchCommentsState?.length === 0 &&
-//     searchTagsState?.length === 0;
+  const emptySearchState =
+    searchProfilesState?.length === 0 &&
+    searchBeamsState?.length === 0 &&
+    // searchReflectionsState?.length === 0 &&
+    searchTagsState?.length === 0;
 
-//   React.useEffect(() => {
-//     if (activeButton !== ButtonValues.ALL) {
-//       analyticsActions.trackEvent({
-//         category: AnalyticsCategories.FILTER_SEARCH,
-//         action: `Filter Search By ${activeButton}`,
-//       });
-//     }
-//     // eslint-disable-next-line react-hooks/exhaustive-deps
-//   }, [activeButton]);
+  React.useEffect(() => {
+    if (activeButton !== ButtonValues.CONTENT) {
+      analyticsActions.trackEvent({
+        category: AnalyticsCategories.FILTER_SEARCH,
+        action: `Filter Search By ${activeButton}`,
+      });
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [activeButton]);
 
-//   const buttonValues = [
-//     {
-//       value: ButtonValues.ALL,
-//       label: t('{{ buttonValueAll }}', { buttonValueAll: ButtonValues.ALL }),
-//     },
-//     {
-//       value: ButtonValues.PEOPLE,
-//       label: t('{{ buttonValuePeople }}', { buttonValuePeople: ButtonValues.PEOPLE }),
-//     },
-//     {
-//       value: ButtonValues.TOPICS,
-//       label: t('{{ buttonValueTopics }}', { buttonValueTopics: ButtonValues.TOPICS }),
-//     },
-//     {
-//       value: ButtonValues.POSTS,
-//       label: t('{{ buttonValuePosts }}', { buttonValuePosts: ButtonValues.POSTS }),
-//     },
-//     {
-//       value: ButtonValues.REPLIES,
-//       label: t('{{ buttonValueReplies }}', { buttonValueReplies: ButtonValues.REPLIES }),
-//     },
-//   ];
+  const buttonValues = [
+    {
+      value: ButtonValues.CONTENT,
+      label: t('{{ buttonValueContent }}', { buttonValueContent: ButtonValues.CONTENT }),
+    },
+    {
+      value: ButtonValues.PEOPLE,
+      label: t('{{ buttonValuePeople }}', { buttonValuePeople: ButtonValues.PEOPLE }),
+    },
+    {
+      value: ButtonValues.TAGS,
+      label: t('{{ buttonValueTags }}', { buttonValueTags: ButtonValues.TAGS }),
+    },
+  ];
 
-//   const onTabClick = (value: ButtonValues) => () => {
-//     setActiveButton(value);
-//   };
+  const onTabClick = (value: ButtonValues) => () => {
+    setActiveButton(value);
+  };
 
-//   const onNavBack = () => {
-//     history.back();
-//   };
+  const onNavBack = () => {
+    history.back();
+  };
 
-//   const searchCount =
-//     searchProfilesState?.length ||
-//     0 + searchPostsState?.length ||
-//     0 + searchCommentsState?.length ||
-//     0 + searchTagsState?.length ||
-//     0;
+  const searchCount =
+    searchProfilesState?.length ||
+    0 + searchBeamsState?.length ||
+    // 0 + searchReflectionsState?.length ||
+    0 + searchTagsState?.length ||
+    0;
 
-//   const allQueriesFinished = React.useMemo(() => {
-//     return (
-//       !searchProfilesReq.isFetching &&
-//       !searchPostsReq.isFetching &&
-//       !searchCommentsReq.isFetching &&
-//       !searchTagsReq.isFetching
-//     );
-//   }, [
-//     searchCommentsReq.isFetching,
-//     searchPostsReq.isFetching,
-//     searchProfilesReq.isFetching,
-//     searchTagsReq.isFetching,
-//   ]);
+  const allQueriesFinished = React.useMemo(() => {
+    return (
+      !searchProfilesReq.isFetching &&
+      !searchBeamsReq.isFetching &&
+      !searchCommentsReq.isFetching &&
+      !searchTagsReq.isFetching
+    );
+  }, [
+    searchCommentsReq.isFetching,
+    searchBeamsReq.isFetching,
+    searchProfilesReq.isFetching,
+    searchTagsReq.isFetching,
+  ]);
 
-//   const isFetchingSearch = React.useMemo(() => {
-//     if (activeButton === ButtonValues.ALL) {
-//       return !allQueriesFinished;
-//     }
-//     return searchKeyword && !searchState[activeButton].done;
-//     // eslint-disable-next-line react-hooks/exhaustive-deps
-//   }, [searchState, activeButton, allQueriesFinished]);
+  const isFetchingSearch = React.useMemo(() => {
+    // if (activeButton === ButtonValues.ALL) {
+    //   return !allQueriesFinished;
+    // }
+    return searchKeyword && !searchState[activeButton].done;
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [searchState, activeButton /* allQueriesFinished */]);
 
-//   return (
-//     <Box data-testid="search-box" fill="horizontal">
-//       <SearchStartCard
-//         searchKeyword={searchKeyword}
-//         handleSearch={handleSearch}
-//         inputPlaceholderLabel={t('Search')}
-//         titleLabel={t('Search')}
-//         introLabel={t('✨ Find what you’re looking for quickly ✨')}
-//         description={t(
-//           'Search everything. Follow wonderful people. And subscribe to any and all topics that get your synapses firing.',
-//         )}
-//       >
-//         <TabsToolbar
-//           noMarginBottom
-//           count={searchCount}
-//           countLabel={t('Results')}
-//           activeButton={activeButton}
-//           tabForm={TAB_TOOLBAR_TYPE.WIDE}
-//           tabButtons={
-//             <>
-//               <StyledSwitchCardButton
-//                 label={t('{{ buttonValuesAll }}', { buttonValuesAll: ButtonValues.ALL })}
-//                 size="large"
-//                 removeBorder={false}
-//                 primary={ButtonValues.ALL === activeButton}
-//                 onClick={onTabClick(ButtonValues.ALL)}
-//               />
-//               <StyledSwitchCardButton
-//                 label={t('{{ buttonValuesPeople }}', { buttonValuesPeople: ButtonValues.PEOPLE })}
-//                 size="large"
-//                 removeBorder={true}
-//                 primary={ButtonValues.PEOPLE === activeButton}
-//                 onClick={onTabClick(ButtonValues.PEOPLE)}
-//               />
-//               <StyledSwitchCardButton
-//                 label={t('{{ buttonValuesTopics }}', { buttonValuesTopics: ButtonValues.TOPICS })}
-//                 size="large"
-//                 removeBorder={true}
-//                 primary={ButtonValues.TOPICS === activeButton}
-//                 onClick={onTabClick(ButtonValues.TOPICS)}
-//               />
-//               <StyledSwitchCardButton
-//                 label={t('{{ buttonValuesPosts }}', { buttonValuesPosts: ButtonValues.POSTS })}
-//                 size="large"
-//                 removeBorder={true}
-//                 primary={ButtonValues.POSTS === activeButton}
-//                 onClick={onTabClick(ButtonValues.POSTS)}
-//               />
-//               <StyledSwitchCardButton
-//                 label={t('{{ buttonValuesReplies }}', {
-//                   buttonValuesReplies: ButtonValues.REPLIES,
-//                 })}
-//                 size="large"
-//                 removeBorder={true}
-//                 primary={ButtonValues.REPLIES === activeButton}
-//                 onClick={onTabClick(ButtonValues.REPLIES)}
-//               />
-//             </>
-//           }
-//           onTabClick={onTabClick}
-//           onIconClick={onNavBack}
-//           hasIcon={true}
-//           hasMobileDesign={true}
-//           buttonValues={buttonValues}
-//           loggedUser={loginState?.pubKey}
-//           style={{ marginBottom: '-1px' }} // overlaps border with parent's bottom border
-//         />
-//       </SearchStartCard>
+  const handleTopMenuClick = () => {
+    return navigateTo?.({
+      appName: '@akashaorg/app-search',
+      getNavigationUrl: () => routes[SETTINGS],
+    });
+  };
 
-//       {allQueriesFinished &&
-//         searchKeyword &&
-//         (isAllTabActive ? emptySearchState : !searchState[activeButton]?.results?.length) && (
-//           <Box margin={{ top: 'medium' }}>
-//             <InfoCard
-//               icon="search"
-//               title={t('No matching results found 👀')}
-//               explanation={t('We could not find any results for your search in Ethereum World.')}
-//               suggestion={t(
-//                 'Make sure you spelled everything correctly or try searching for something else.',
-//               )}
-//             />
-//           </Box>
-//         )}
+  return (
+    <Box customStyle="flex(& col)">
+      <SearchStartCard
+        searchKeyword={searchKeyword}
+        handleSearch={handleSearch}
+        inputPlaceholderLabel={t('Search')}
+        // titleLabel={t('Search')}
+        // introLabel={t('✨ Find what you’re looking for quickly ✨')}
+        // description={t(
+        //   'Search everything. Follow wonderful people. And subscribe to any and all topics that get your synapses firing.',
+        // )}
+        handleTopMenuClick={handleTopMenuClick}
+      >
+        <SwitchCard
+          activeButton={activeButton}
+          onTabClick={onTabClick}
+          onIconClick={onNavBack}
+          buttonValues={buttonValues}
+          loggedUser={loginState?.pubKey}
+          style="-mb-px" // overlaps border with parent's bottom border
+        />
+      </SearchStartCard>
+      {activeButton === ButtonValues.CONTENT && (
+        <SearchAppFilter
+          dropdownMenuItems={dropdownMenuItems}
+          selected={selected}
+          setSelected={setSelected}
+        />
+      )}
+      {searchKeyword === '' && <SearchDefaultCard />}
 
-//       <Box margin={{ top: 'medium' }}>
-//         {(activeButton === ButtonValues.ALL || activeButton === ButtonValues.PEOPLE) &&
-//           searchProfilesState?.map((profileData: IProfileData, index: number) => (
-//             <Box key={index} pad={{ bottom: 'medium' }}>
-//               <ProfileSearchCard
-//                 handleFollow={() => handleFollowProfile(profileData.pubKey)}
-//                 handleUnfollow={() => handleUnfollowProfile(profileData.pubKey)}
-//                 isFollowing={followedProfiles.includes(profileData?.pubKey)}
-//                 loggedEthAddress={loginState?.ethAddress}
-//                 profileData={profileData}
-//                 followLabel={t('Follow')}
-//                 unfollowLabel={t('Unfollow')}
-//                 descriptionLabel={t('About me')}
-//                 followingLabel={t('Following')}
-//                 followersLabel={t('Followers')}
-//                 postsLabel={t('Posts')}
-//                 shareProfileLabel={t('Share')}
-//                 profileAnchorLink={'/profile'}
-//                 onClickProfile={() => handleProfileClick(profileData.pubKey)}
-//               />
-//             </Box>
-//           ))}
+      {
+        /* allQueriesFinished */ !isFetchingSearch &&
+          searchKeyword &&
+          /* isAllTabActive ? emptySearchState : */ !searchState[activeButton]?.results?.length && (
+            <Box customStyle="mt-4">
+              <InfoCard
+                explanation={t('Oops! Looks like there’s no results for the word')}
+                keyword={searchKeyword}
+                preposition_in={t('in')}
+                section={activeButton}
+                suggestion={t('Try searching for something else or try a different Category!')}
+              />
+            </Box>
+          )
+      }
 
-//         {(activeButton === ButtonValues.ALL || activeButton === ButtonValues.TOPICS) &&
-//           searchTagsState?.map((tag: ITag, index: number) => (
-//             <Box key={index} pad={{ bottom: 'medium' }}>
-//               <TagSearchCard
-//                 tag={tag}
-//                 subscribedTags={tagSubscriptionsState}
-//                 subscribeLabel={t('Subscribe')}
-//                 unsubscribeLabel={t('Unsubscribe')}
-//                 tagAnchorLink={'/@akashaorg/app-akasha-integration/tags'}
-//                 onClickTag={() => handleTagClick(tag.name)}
-//                 handleSubscribeTag={handleTagSubscribe(true)}
-//                 handleUnsubscribeTag={handleTagSubscribe(false)}
-//               />
-//             </Box>
-//           ))}
-//         {(activeButton === ButtonValues.ALL || activeButton === ButtonValues.POSTS) &&
-//           searchPostsState?.map(itemData => (
-//             <EntryCardRenderer
-//               key={itemData.itemId}
-//               itemData={itemData}
-//               itemType={EntityTypes.POST}
-//               logger={props.logger}
-//               singleSpa={singleSpa}
-//               ethAddress={loginState?.ethAddress}
-//               onContentClick={handleEntryNavigation}
-//               navigateTo={navigateTo}
-//               onRepost={handleRepost}
-//               onAvatarClick={handleAvatarClick}
-//               onMentionClick={handleMentionClick}
-//               onTagClick={handleTagClick}
-//               contentClickable={true}
-//               locale={locale}
-//               moderatedContentLabel={t('This content has been moderated')}
-//               ctaLabel={t('See it anyway')}
-//               uiEvents={props.uiEvents}
-//               navigateToModal={props.navigateToModal}
-//             />
-//           ))}
-//         {(activeButton === ButtonValues.ALL || activeButton === ButtonValues.REPLIES) &&
-//           searchCommentsState?.map(itemData => (
-//             <EntryCardRenderer
-//               key={itemData.itemId}
-//               itemData={itemData}
-//               itemType={EntityTypes.REPLY}
-//               logger={props.logger}
-//               singleSpa={singleSpa}
-//               ethAddress={loginState?.ethAddress}
-//               onContentClick={handleEntryNavigation}
-//               navigateTo={navigateTo}
-//               onRepost={handleRepost}
-//               onAvatarClick={handleAvatarClick}
-//               onMentionClick={handleMentionClick}
-//               onTagClick={handleTagClick}
-//               contentClickable={true}
-//               locale={locale}
-//               moderatedContentLabel={t('This content has been moderated')}
-//               ctaLabel={t('See it anyway')}
-//               uiEvents={props.uiEvents}
-//               navigateToModal={props.navigateToModal}
-//             />
-//           ))}
-//       </Box>
-//       {isFetchingSearch && (
-//         <Box pad="large">
-//           <Spinner />
-//         </Box>
-//       )}
-//       {/* triggers intersection observer */}
-//       <Box pad="xxsmall" ref={loadmoreRef} />
-//     </Box>
-//   );
-// };
+      <Box customStyle="mt-4">
+        {activeButton === ButtonValues.PEOPLE &&
+          searchState[ButtonValues.PEOPLE].done &&
+          !!searchProfilesState.length && (
+            <>
+              {
+                <SearchResultCount
+                  countLabel={t('Found {{count}} results for {{searchKeyword}} in {{category}}', {
+                    count: 0 + searchProfilesState?.length,
+                    searchKeyword: searchKeyword,
+                    category: activeButton,
+                  })}
+                />
+              }
+              {searchProfilesState?.map((profileData: IProfileData, index: number) => (
+                <Box key={index} customStyle="pb-4">
+                  <ProfileSearchCard
+                    handleFollow={() => handleFollowProfile(profileData.pubKey)}
+                    handleUnfollow={() => handleUnfollowProfile(profileData.pubKey)}
+                    isFollowing={followedProfiles.includes(profileData?.pubKey)}
+                    loggedEthAddress={loginState?.ethAddress}
+                    profileData={profileData}
+                    followLabel={t('Follow')}
+                    unfollowLabel={t('Unfollow')}
+                    descriptionLabel={t('About me')}
+                    followingLabel={t('Following')}
+                    followersLabel={t('Followers')}
+                    beamsLabel={t('Beams')}
+                    shareProfileLabel={t('Share')}
+                    profileAnchorLink={'/profile'}
+                    onClickProfile={() => handleProfileClick(profileData.pubKey)}
+                  />
+                </Box>
+              ))}
+            </>
+          )}
+        {activeButton === ButtonValues.TAGS &&
+          searchState[ButtonValues.TAGS].done &&
+          !!searchTagsState.length && (
+            <>
+              {
+                <SearchResultCount
+                  countLabel={t('Found {{count}} results for {{searchKeyword}} in {{category}}', {
+                    count: 0 + searchTagsState?.length,
+                    searchKeyword: searchKeyword,
+                    category: activeButton,
+                  })}
+                />
+              }
+              <BasicCardBox customStyle="pb-0">
+                {searchTagsState?.map((tag: ITag, index: number) => (
+                  <Box key={index}>
+                    <TagSearchCard
+                      tag={tag}
+                      subscribedTags={tagSubscriptionsState}
+                      subscribeLabel={t('Subscribe')}
+                      unsubscribeLabel={t('Unsubscribe')}
+                      tagAnchorLink={'/@akashaorg/app-akasha-integration/tags'}
+                      onClickTag={() => handleTagClick(tag.name)}
+                      handleSubscribeTag={handleTagSubscribe(true)}
+                      handleUnsubscribeTag={handleTagSubscribe(false)}
+                    />
+                    {index < searchTagsState?.length - 1 && <Divider />}
+                  </Box>
+                ))}
+              </BasicCardBox>
+            </>
+          )}
+        {activeButton === ButtonValues.CONTENT &&
+          searchState[ButtonValues.CONTENT].done &&
+          !!searchBeamsState.length && (
+            <>
+              {
+                <SearchResultCount
+                  countLabel={t('Found {{count}} results for {{searchKeyword}} in {{category}}', {
+                    count: 0 + searchBeamsState?.length,
+                    searchKeyword: searchKeyword,
+                    category: activeButton,
+                  })}
+                />
+              }
+              {searchBeamsState?.map(itemData => (
+                <EntryCardRenderer
+                  key={itemData.itemId}
+                  itemData={itemData}
+                  itemType={EntityTypes.POST}
+                  logger={props.logger}
+                  singleSpa={singleSpa}
+                  ethAddress={loginState?.ethAddress}
+                  onContentClick={handleEntryNavigation}
+                  navigateTo={navigateTo}
+                  onRebeam={handleRebeam}
+                  onAvatarClick={handleAvatarClick}
+                  onMentionClick={handleMentionClick}
+                  onTagClick={handleTagClick}
+                  contentClickable={true}
+                  locale={locale}
+                  moderatedContentLabel={t('This content has been moderated')}
+                  ctaLabel={t('See it anyway')}
+                  uiEvents={props.uiEvents}
+                  navigateToModal={props.navigateToModal}
+                />
+              ))}
+            </>
+          )}
+        {/* {(activeButton === ButtonValues.ALL || activeButton === ButtonValues.REPLIES) &&
+          searchReflectionsState?.map(itemData => (
+            <EntryCardRenderer
+              key={itemData.itemId}
+              itemData={itemData}
+              itemType={EntityTypes.REPLY}
+              logger={props.logger}
+              singleSpa={singleSpa}
+              ethAddress={loginState?.ethAddress}
+              onContentClick={handleEntryNavigation}
+              navigateTo={navigateTo}
+              onRepost={handleRebeam}
+              onAvatarClick={handleAvatarClick}
+              onMentionClick={handleMentionClick}
+              onTagClick={handleTagClick}
+              contentClickable={true}
+              locale={locale}
+              moderatedContentLabel={t('This content has been moderated')}
+              ctaLabel={t('See it anyway')}
+              uiEvents={props.uiEvents}
+              navigateToModal={props.navigateToModal}
+            />
+          ))} */}
+      </Box>
+      {isFetchingSearch && (
+        <Box customStyle="p-8 flex flex-col items-center justify-center m-auto space-y-8">
+          <SeventyFivePercentSpinner
+            color={{ light: 'secondaryLight', dark: 'secondaryDark' }}
+            size="xxl"
+          />
+          <Text variant="footnotes2">Searching...</Text>
+        </Box>
+      )}
+      {/* triggers intersection observer */}
+      <Box customStyle="p-2" ref={loadmoreRef} />
+    </Box>
+  );
+};
 
-// export default React.memo(SearchPage);
+export default React.memo(SearchPage);
